Changes address parsing to be aware of the TCL's '+'-notation.
The TCL transport is more like an intermediate layer that itself uses
other transports to connect to other nodes. To determine whether or not
the TCL transport should be used and which "subtransport" should be used
to connect to the denoted peer, the TCL uses a '+'-notation in the
protocol-fragment of the address string, e.g. "tcl+tcp://". Hence, the
address parsing routine needs to be patched accordingly, such that the
TCL transport gets invoked whenever the address starts with "tcl+" and
gets the remainder of the address to determine which subtransport it is
supposed to use.